Update the Antora project configuration
This makes a minor changes the directory, inside the deployment
directory, where the project files are deployed. Previously, it was set
to "owncloud-docs", but for some reason it's been changed to "server".
This breaks all of the static file links within the documentation. To
avoid a lot of work correcting the links, the directory name's being
changed in antora.yml instead.