Add the Asciidoc Live Preview Plugin to the guide
I don't have an issue with installing Antora's build tools, but then
it's pretty smooth on Linux. However, I get that it may not be for
everyone — especially if they're working on Windows. So, I've updated
the getting started guide to cover the Asciidoc Live Preview Plugin, in
case they're looking for the simplest way to get started editing
AsciiDoc text.