Use Courier and Times New Roman custom fonts
While the built-in files are fine, for the most part, they have problems
with rendering the code examples and some other text, as the character
set of the default fonts doesn't support more than Win-ASCII. Given
that, custom fonts need to be used if the manuals are to render
correctly. I've used Courier and Times New Roman, as they match the
closest, however, they're not freely distributable. Given that, they're
going to have to be included before the script can be run.