Create a Makefile to automate creation of the PDF manuals
As there's quite a bit going on when generating the PDF manuals from the
source files, I created a Makefile to make the process automatable,
repeatable, and to take the hassle out of it. It has variables for all
of the key settings, defines a minimum number of targets to get the work
done and to clean up after itself, and it breaks the functionality down
into meaningful functions (one's not yet ready for primetime). What it
doesn't do is to check if asciidoctor-pdf is available. That will be
added a little bit later.