Move the administration manual to a separate, sub-module
There's been some discussion about making the administration manual a
separate, sub-module in the docs recently. While I was initially against
the idea, after doing so, and setting the default ROOT module as an
overview module, it makes perfect sense. This is for several reasons;
firstly, it's easier to navigate when looking at the physical file
system, and secondly, because the ROOT module can provide an overview or
introduction to the sub-module contents.