Escape AsciiDoc attributes as required
Some of the code samples contain what AsciiDoc interprets as attributes.
As a result, since they're not attributes, and therefor undeclared, when
the documentation is built, Asciidoctor throws warnings. Given that,
they have to be escaped, where necessary. Specifically, they can be used
in code blocks, but not inline code statements.