Improve the getting started guide
As it was, the guide wasn't near as helpful as I'd have liked it to have
been. So, I've created this patch to make it a lot clearer as to what's
required, while attempting to not assume that the user has much prior
experience, yet not come across as condescending at the same time. This
patch draws heavily from the Zend Framework contributing guide, which
can be found at
https://github.com/zendframework/maintainers/blob/master/CONTRIBUTORS.md.