Move build documentation instructions to the build-the-docs file
While it was helpful to have the information in the README, it would
likely cause confusion now that the getting started guide's been updated
and better links to the other "build-the-docs" guide. Given that, the
information's been moved from the README and added to the build the docs
guide, with a link added. Now, there's one source of truth.