Create a custom Dockerfile
While, normally, the official Antora Docker image can be used to build
Antora documentation, as the ownCloud setup integrates the Lunr site
search, it needs some extra packages, and doesn't build properly without
including custom packages. Given that, the package.* files needed to be
included in the image so that the identified packages can be installed
and available to the container. And, as these things normally have it,
to run "npm i" extra system packages needed to be installed as well. So,
this handles all of those needs.