Add a site.url to the Playbook
This is required so that, when being served, Lunr knows where to look to
index the documentation content. I've decided to leave it set as
http://localhost:5000, which I use locally, because I'm guessing that,
this might work as part of the build pipeline, and that, after the
search index file's generated, that sed could be used to replace
references to localhost:5000 to the host domain.