Split out reusable content into include files
The documentation on enabling the encryption app and the encryption
types comes up several times throughout the encryption documentation.
So, it makes sense to split it out into include files and include it
where required, rather than to duplicate the content "all over the
place". That way, the content is far less costly to maintain.