Update the production playbook file
After coming to understand Antora more over the last few days, it seems
important to make several changes to the Playbook files:

1. Create Playbook files specific to each environment, such as
local/dev, testing & staging (if required in the future), and
production.
2. Don't hardcode the ui.bundle.url link. This can be
specified/overridden on the command line, and likely will be as part of
the CI/CD pipeline.

Given that, the file's been renamed from site.yml to site.prod.yml, and
the ui.bundle.url setting's been removed.