Add package.json and package-lock.json (#216)
After a lot of thought, and some troubleshooting, it makes sense (to me at least, if no one else) to install all required packages locally for building the documentation. To simplify that process, this commit provides an initial package.json and package-lock.json file, that contain the core dependencies, along with other package metadata.