Implement site search (#189)
* Add a custom site generator which integrates Lunr

This custom generator ensures that a Lunr index will be generated, when
the documentation is generated.

* Create a custom Dockerfile

While, normally, the official Antora Docker image can be used to build
Antora documentation, as the ownCloud setup integrates the Lunr site
search, it needs some extra packages, and doesn't build properly without
including custom packages. Given that, the package.* files needed to be
included in the image so that the identified packages can be installed
and available to the container. And, as these things normally have it,
to run "npm i" extra system packages needed to be installed as well. So,
this handles all of those needs.

* Add a site.url to the Playbook

This is required so that, when being served, Lunr knows where to look to
index the documentation content. I've decided to leave it set as
http://localhost:5000, which I use locally, because I'm guessing that,
this might work as part of the build pipeline, and that, after the
search index file's generated, that sed could be used to replace
references to localhost:5000 to the host domain.

* Update the Playbook's UI bundle URL

This change points to the latest version of the UI theme, which
integrates site search with Lunr.

* Update the build the docs documentation

This change updates the instructions that show how to build the
documentation. Technically, at the moment, the Docker container
instructions won't work, as there's some issue with the container that
I'm trying to resolve. But the version that references the locally
installed tools does work. Given that, I'm leaving it for the moment and
will make necessary changes to the relevant files just as soon as I know
more.

* Replace hardcoded site.url with an interpolatable variable

@tboerger rightly pointed out in
https://github.com/owncloud/docs/pull/189#pullrequestreview-172376664,
that hard-coding site.url isn't the correct approach. Given that, this
commit replaces it with an interpolatable string.

* Correct package.json

Ensure that all the necessary dependencies are included and sort the
contents so that it's easier to read (and maintain).

* Update the build documentation