Simplify the XSS link in the developer security documentation
While there's nothing, necessarily, wrong with the link as it was, it's
at risk of becoming hard to read. However, if a page attribute is used,
then the link can be written as normal, with no unnecessary extra
formatting, and also makes the prose easier to read, as there's no infix
URL to scan.