Add a make target for running an Antora Xref validator (#220)
This is designed as another step for simplifying working with Antora
projects. It uses a custom Antora build pipeline, that makes use of a
custom project generator, that can validate all Xrefs in a Playbook's
source files and output a report based on what it finds, if there are
any invalid Xrefs found.