Remove merge redirects script
After a bit of thought on this, it makes more sense (to me, if to no one
else) to convert the _redirects file to a netlify.toml file. It will be
merged with the version created in #323 and shares the same benefits and
motivations.  Secondly, the merge script's not necessary if static
redirect pages are generated instead of creating a Netlify _redirects
file that has to be merged with an Antora generated version.