Update redirect rules
This commit should update the redirect rules to cover all versions of
the manuals, prior to the latest version, regardless of domain, to the
yet to be created, "old docs" domain. Any requests to the latest docs in
the original docs will be redirected to the correct location in the new
structure.