Clean up the Let's Encrypt Apache documentation
This is part of the process of fixing up / updating this section of the
documentation. It reflows the sentences so that they're easier to
maintain over the course of time, and extracts the URLs out into page
attributes. This makes the text easier to read (it doesn't break the
flow) and allows the URLs to also be reusable.