Update the issue template to make it simpler
This change helps ensure that if someone is in a hurry, they can write
what they need to as quickly as possible, and not have to fill out
unnecessary forms.

It also:

- Reorders the existing items top-loading them from the most to the
least important,
- Shows that some items are optional,
- Shows that some items are for required by the docs team, not for
people creating issues.

I hope this balances the requirements for people who are in a hurry, and
those who have more time.