Revise update-search drone task
This change should fix the update script so that it won't break at the
end of the process, when it's asked whether it wants to update the
nb_hits value in the config file. How to deal with the change in value
wrt git, that I'm not sure about, yet.