Mark UI Bundle as a snapshot
As the UI bundle changes from time to time, it needs to be marked as a
snapshot, so that Antora will download it and get the latest changes.
Not having this option set has lead to several UI updates, including the
integrated site search not being made available.

See https://docs.antora.org/antora/2.0/playbook/configure-ui/#snapshot
for more information about this setting.