Update the Drone build-html pipeline configuration
From what I've been able to piece together, if a secret is to be used,
it has to be explicit in the configuration; it can't be implicit. So two
secrets missing from build-html have been added. Secondly, the
algolia_app_id secret mapping's been removed, as there's no reference to
the mapped key name.