Update the legal settings documentation
This change reworks the information in the legal settings documentation
so that it is easier to read, adds a link to the details about how to,
manually, edit email templates to add links to the imprint and privacy
urls, and to show where the information is displayed by default.
This fixes #572.