Reworked the session management content slightly
I don't want to force my own ideas here, just want to provide assistance
and guidance. What I've done is to use some of Antora's experimental
attributes to make suggestions for using menus and buttons clearer,
streamlining the text, and to extend the content to reflect recent
changes to the respective sections of functionality.