Globally enable the experimental attribute
This fixes #975. This is an attribute that we've been testing for some
time to see if it enhances readability. We believe that it is ready to
be used throughout the documentation. So, the attribute's being enabled
globally with this change, and any per/page references to it have been
removed.  For more information, please see
https://docs.antora.org/antora/2.0/asciidoc/ui-macros/#set-the-experimental-attribute.