Link All Category Roots to Index Pages (#1325) (#1349)
* Update content structure

As part of the process of adding links to all section headers, some of
the content needed to be restructured. This change implements the
restructure and, quite importantly, creates page aliases to all the
pages that were moved, ensuring that there are no breaks in user
experience.

* Add index pages for all (missing) sections

This change uses a partial to simplify the process of creating index
pages for all the missing section pages. The partial, effectively,
provides a reusable template that can be filled in as necessary, saving
time creating and maintaining each index page.

* Update navigation links

This change links each section header to the newly created landing
pages.

* Update the help and communication page

The page needed a slight restructure to fit in with the new link
structure. It also received a bit extra work to ensure that it followed
the style guide.

* Removed unrequired file