[WIP] Rework the Anti-Virus documentation (#1238)
* Rework the Anti-Virus documentation

This addresses the majority of the changes requested in #333, and makes
a series of other, relevant, changes. It reflows the text, fixes up
spelling and grammar mistakes, simplifies the instructions and
reorganises them to be easier to read. This fixes #333.

* Minor text updates

* Further updated the anti-virus documentation

Used more of the experimental features to draw greater attention to the
configuration form elements and moved the configuration warnings section
after the configuration settings, as the warnings are supplemental to
the configuration settings.

* Document that ClamAV can be configured using occ

* Add missing source block meta configuration