Remove the list of included files in the book configuration files
This change reduces the book configuration files to just the header
information as, in a later commit, the pages to be included, which will
make up the content of the PDF file, will be auto-generated later. This
approach helps ensure that the PDF file content is automatically kept in
sync with any changes in the Antora source files.