Refactor the PDF generation process
Based on a suggestion from @tboerger, this commit finishes the refactor
of the PDF generation process, so that it's simpler and easier to
maintain. It uses a combination of PHP and shell scripts, along with
changes to the PDF configuration files to help ensure that the process
is as simple and maintainable as possible.