Fix table formatting
It turns out that asciidoctor-pdf throws a Prawn::Errors::CannotFit,
fatal, error if a table has some, but not all, column widths specified.
This commit sets a column width value for all columns in the table that
didn't have them set, and makes some other, minor table fixes as well.