Fix broken PDF generation and docs build process (#1357)
* Remove custom section ids

While they can be handy, from observation, they’re often unnecessary.
Looking through around 80% of the ones that have been removed, they're
exactly what would be auto-generated by Antora. Given that, they
increase maintenance overhead for no benefit. After removing them all,
and running yarn validate, there are no broken links.

* Update the book (PDF) configuration files

This change updates the book configuration files for the administration
and developer manuals, as they weren't updated when the structure of the
core documentation was changed.

* Revert to the older partialsdir, examplesdir, & imagesdir include syntax

The newer partial$, example$, and image$ doesn't work when building PDFs
using asciidoctor-pdf, as it doesn't understand the syntax. Given that,
the older format is, temporarily, being reverted to, until a better
solution can be found.

* Add the partial to the user manual

Sadly, at least for the time being, asciidoctor-pdf doesn't understand
the new partial, image, and example include syntax, so files cannot be
included from other modules. As a result, the partial that was
referenced from the Admin module now has to be included from the
including module. :-(

* Remove the list of included files in the book configuration files

This change reduces the book configuration files to just the header
information as, in a later commit, the pages to be included, which will
make up the content of the PDF file, will be auto-generated later. This
approach helps ensure that the PDF file content is automatically kept in
sync with any changes in the Antora source files.

* Revert to the old include syntax

This helps ensure that asciidoctor-pdf can generate the correct include
paths, whether that's for image, partial, or example files.

* Add a PHP library for generating the PDF config file include refs

This change includes the
settermjd/antora-nav-to-asciidoctor-pdf-config-line-converter package
for handling the generation of asciidoctor-pdf config file include ref
generation based on the supplied Antora navigation file.

* Refactor the PDF generation process

Based on a suggestion from @tboerger, this commit finishes the refactor
of the PDF generation process, so that it's simpler and easier to
maintain. It uses a combination of PHP and shell scripts, along with
changes to the PDF configuration files to help ensure that the process
is as simple and maintainable as possible.

* Ignore the PHP Composer vendor directory

* Rewrite the Makefile as a shell script

After much discussion, the Makefile has been rewritten as a shell
script. While making that change, the recently introduced PHP script was
also removed, and its functionality rolled into the new shell script.

* Update Drone configuration for building PDF files

* Make the cli script more robust

* Add ios-app to content sources

* Remove limit impersonation animated gif image

An animated gif won't work in PDF format.

* Fix table formatting

It turns out that asciidoctor-pdf throws a Prawn::Errors::CannotFit,
fatal, error if a table has some, but not all, column widths specified.
This commit sets a column width value for all columns in the table that
didn't have them set, and makes some other, minor table fixes as well.

* Remove unnecessary mkdir check

As @tboerger clarified, the check isn't required when mkdir uses the -p
option.