Fix PDF generation process in 10.1 (#1378)
* Fix table formatting

It turns out that asciidoctor-pdf throws a Prawn::Errors::CannotFit,
fatal, error if a table has some, but not all, column widths specified.
This commit sets a column width value for all columns in the table that
didn't have them set, and makes some other, minor table fixes as well.

* Migrate to a shell-script based build tool

This copies the relevant changes from 10.2, in that it copies the build
script, removes the Makefile, and updates the Drone configuration.

* Migrate Antora include syntax to the old AsciiDoc syntax

This change is being made as asciidoctor-pdf doesn’t understand the
Antora syntax, so include files are not found, effectively, breaking
the generated PDF.

* Add ios-app to content sources

* Remove the list of included files in the book configuration files

This change reduces the book configuration files to just the header
information as, in a later commit, the pages to be included, which will
make up the content of the PDF file, will be auto-generated later. This
approach helps ensure that the PDF file content is automatically kept in
sync with any changes in the Antora source files.