redis address correction (#1437)
* redis address correction

the ip address in the example was machine specific. replaced it with the loop back address.

* Use site attributes in the Redis code example

This change replaces the hard strings with site attributes to help
ensure example consistency throughout the documentation. It also ensures
that Redis is written properly as well. For more information on Antora
site attributes, please refer to
https://docs.antora.org/antora/2.0/page/page-and-site-attributes/.

* Use site attributes in the Redis code example

This change replaces the hard strings with site attributes to help
ensure example consistency throughout the documentation. It also ensures
that Redis is written properly as well. For more information on Antora
site attributes, please refer to
https://docs.antora.org/antora/2.0/page/page-and-site-attributes/.