Document new options for adding users via provisioning API and occ
This change, documents that when creating users, either one or both of a
username and password can be supplied. However, if one is supplied, the
other isn't. More specifically; if an email address is supplied and a
password isn't, that a password will be automatically generated and the
user will be emailed a link to change their password.