Update the OCS Share API Get Shares From Specific File or Folder docs (#1570)
* Update the OCS Share API Get Shares From Specific File or Folder docs

This change updates the request attributes table for the OCS Share API
get shares from specific file or folder documentation. Specifically, it
adds the shared_with_me, format, and state attributes; two of which are
required for the pending shares API in oC 10.0.9+. This change,
therefore, relates to #1193.

* Document the new OCS endpoints for accepting and rejecting shares

This documents the above endpoints, and adds code examples in curl, PHP,
Ruby, and Go.