Update unit test documentation
This change fixes #320 by fleshing out the details around running unit
tests in ownCloud apps. It clarifies that core apps are tested when
testing core, and that other apps can be tested separately.

As a large number of apps use the Makefile for automating tests, and
that it's referenced earlier in the file, then it makes sense to link to
it here, so that users can use it, should they choose to.