Guests app documentation
Added documentation here for the guests app as we had not have anything before that except one paragraph in user roles. But since our users / customer use the guests app intensively we should have a better documentation about it. This should be a start, we can add things to it later on as we discover possible use cases and issues.