Update references to totp user is not found response examples
The include link to the JSON example was corrected and the XML example
file was updated to keep consistency with the JSON example as well as
the section. The last thing that I want to have happen is to have
confusion at some future point as to whether the XML file should be
there, is incorrectly linked, etc, all because of a minor name
difference.