Left align PDF content
Dusan Cirkovic made a suggestion (see #680) that some content in the PDF
was not formatted (aligned) correctly. The reason for the seemingly
broken formatting is because the content is justified, and not left
aligned. While it works well in a lot of cases, there are some where it
doesn't. And, after reviewing the PDF with left aligned text, it looks
far better.