prefer remote.php/webdav route remote.php/dav/files/... (#1822)
* prefer remote.php/webdav route remote.php/dav/files/...

Gnome file managers truncate webdav urls after two path components

* Update access_webdav.adoc

In Ubuntu 18.04 we have:

```
-rw------- 1 root root 2648 Apr 27  2016 /etc/davfs2/secrets
```

But we want `~/.davfs2/secrets` owned by the user. 
Use a combination is sudo and cat instead of cp to get that right.

* Update access_webdav.adoc

Entries on `/etc/fstab` need a scheme.
Entries in `~/.davfs2/secrets` should refer to the mount point, not to the URL.
Show sudo with those commands that need root. 
Operations in the user's home should not be done as root.