Update cli script (#1595)
* Add dry-run functionality to building the PDF manuals

When used, the dry-run functionality will output the content which the 
PDF manuals will be built from. This can be used to debug the content,
to ensure that it’s what is expected.

* Add ability to build a specific manual

This change allows the user (or build tool) to choose a specific manual
to build, instead of building all three. By default, all three will be
built, which ensures backwards compatibility with the previous version.
This relates to #1568.