Minor grammatical and formatting revisions (#1870)
The text was changed in two, token, ways, so that it was that much
clearer in its intent. The NOTE admonition at the top of the page was
changed so that it now only contains the first two sentences. This is
because the remaining sentences originally contained in it read as a
normal part of the prose on that page, and don't require special
attention.