Fixed a bug in the cli script when generating PDF files (#1883)
For some reason, which is currently unknown to me, I inserted a call to the validate_manual function in the build_manuals function. I'm not sure why I did that, but it broke PDF generation for all manuals (but not for individual manuals). This change removes the call, adds a bit more error checking for the call to getopts, to make it more robust, and also makes use of the "strict mode", to also make the script more robust, in the right way. For more information on the "strict mode", check out http://redsymbol.net/articles/unofficial-bash-strict-mode/.