Fix build errors (#1754) (#1925)
* Fix up incorrect header levels

Some of the headers were at inconsistent nesting levels within the
structure of the page that they were one, others were not included with
the correct level offset. Both of these situations are corrected in this
change.

* Update references to totp user is not found response examples

The include link to the JSON example was corrected and the XML example
file was updated to keep consistency with the JSON example as well as
the section. The last thing that I want to have happen is to have
confusion at some future point as to whether the XML file should be
there, is incorrectly linked, etc, all because of a minor name
difference.

* Fix missing attribute reference warnings

There were a number of places where either the attribute wasn't defined,
wasn't being passed on to an included file, or was being incorrectly
interpreted as an attribute. This change corrects all of these problems.

* Remove unnecessary custom ids

These aren't necessary, as ones exactly the same will be generated by
Antora.