Fix (potential) asciidoctor shell script bug (#1891)
While I've never seen it before on macOS and haven't seen the bug
reported in Drone, when running bin/cli on Linux (Mint 19.2) today, I
received the following error:

asciidoctor: FAILED: input path /dev/fd/63 is a fifo, not a file

According to the asciidoctor man page:

The asciidoctor synopsis is: "asciidoctor [OPTION]... FILE..."
Secondly: "If FILE is - then the AsciiDoc source is read from standard
input."

From that, I can only determine that it should, on all platforms, have
problems with anonymous pipes. Strange why it didn't turn up sooner.

Making the change in this commit correct the issue (on Linux) and also
turned up a few other bugs in the documentation (that are being
corrected in other PRs).