Expand the note about NGINX not being officially supported
The redirect that I added previously didn't really help the user and may
cause more confusion than assistance; as there is no warning that it
will happen, nor why it has happened. By expanding the note, at least
the use can choose for themselves if they want to read the Apache
documentation, and have a bit more information to work with.