Correct level offset values for included pages
The level offsets are being lowered as they are making the section
headers be generated out of sequence. I think they were set to the
original value as that is the common value. However, it's not correct in
this case, as leveloffset values are relative to where the file is being
included, and not at some top-level. This relates to #2045.