Update the navigation files to fix the navigation bug in the PDF files
Given how the navigation file (nav.adoc) were structured, the first
top-level page was not nested as deeply as the rest. This resulted in a
potentially confusing navigation structure in the PDF files (but didn't
in the HTML manuals). So, this change moves the first page in most of
the manuals to a level of indentation equal to the other top-level
navigation items.