Move the --database configuration values next to its reference
When working through setting up PostgreSQL with ownCloud, I didn't find
the configuration value and went on a hunt through the server source and
the documentation to find the information. On updating this section, I
found the information at the bottom of page. Having the option and
values next to each other should avoid anyone else going on a long hunt
in the future.