Move Ubuntu 18.04 guide to a new guides directory
This change is motivated by the upcoming creation of several similar
guides to the one for Ubuntu for RHEL and SLES. Before, when there was
only one guide, it was logical enough to not include it in a guides
directory. But now that there will be several, it makes sense to
reorganise it.