Possible fix for owncloud/core#38074
OWNCLOUD_DOMAIN can have a port number, and should have one, if non-standard.
OWNCLOUD_DOMAIN must not have a protocol, though.

Not sure if OWNCLOUD_PORT should be implicitly applied to OWNCLOUD_DOMAIN too, aparently it is not.
That would decide, if this an issue with docu or with code.

Backports needed: 10.5, 10.6 ... maybe more?