Update issue_template.md
I removed the comments. I find it a challenge for new users to have them read through all of the text just to create a quick issue.

When you find something and you want to create an issue to mark your findings for later, you don't want to lose what is in your head by reading the commentary of the needed information, you want to quickly fill in what you have in the pre defined mould. The better we make the mould the better will be the issues. The problem is not to write too much and not too little.

The current issue template was too much from my view. I often saw my support people halting and reading through all of the comments and not being sure what to do.