Update ubuntu_20_04.adoc
Restarting/reloading the Apache service in each step is redundant. The user does in "Change the Document Root,"  "Enable the Virtual Host Configuration," and "Enable the Recommended Apache Modules."

It is much better to do it once after the Apache configuration.

The Element `echo "Enabling Apache Modules"`
shows the message into the console and does not interact with the installation process.