revert #4381
Variable names have been changed, this causes errors / warnings doing the setup. 

You will get the following warnings:

WARNING: The ADMIN_USERNAME variable is not set. Defaulting to a blank string.
WARNING: The ADMIN_PASSWORD variable is not set. Defaulting to a blank string.

Those 2 Variables are being set to be the values of other variables in the yaml file.

If you remove / rename them - the variables aren't being set.