Update site.yml
A new common attribute was added in docs-server - `oc-complete-base-url: 'https://download.owncloud.com/server/stable'` needs to be added in the global section of site.yml in docs as this is the assembling branch. If this is not added here, it cant be rendered resolved and is not rendered properly.